remove top level module export and refactor conftests (#48)
