+++ space2stats_api/src/tests/test_api.py
@@ -1,8 +1,3 @@
-import psycopg
-from fastapi.testclient import TestClient
-from pytest import fixture
-from pytest_postgresql.janitor import DatabaseJanitor
-
 aoi = {
     "type": "Feature",
     "geometry": {
@@ -19,49 +14,6 @@ aoi = {
     },
     "properties": {},
 }
-
-
-@fixture(scope="session")
-def database(postgresql_proc):
-    janitor = DatabaseJanitor(
-        user=postgresql_proc.user,
-        host=postgresql_proc.host,
-        port=postgresql_proc.port,
-        dbname=postgresql_proc.dbname,
-        version=postgresql_proc.version,
-        password=postgresql_proc.password,
-    )
-    janitor.init()
-    yield postgresql_proc
-    janitor.drop()
-
-
-@fixture(scope="session")
-def database_url(database):
-    db_url = f"postgresql://{database.user}:{database.password}@{database.host}:{database.port}/{database.dbname}"
-    with psycopg.connect(db_url, autocommit=True) as conn:
-        with conn.cursor() as cur:
-            cur.execute("""
-                CREATE TABLE IF NOT EXISTS space2stats (
-                    hex_id TEXT PRIMARY KEY,
-                    sum_pop_2020 INT,
-                    sum_pop_f_10_2020 INT
-                );
-            """)
-            cur.execute("""
-                INSERT INTO space2stats (hex_id, sum_pop_2020, sum_pop_f_10_2020)
-                VALUES ('hex_1', 100, 200), ('hex_2', 150, 250);
-            """)
-    return db_url
-
-
-@fixture(autouse=True)
-def client(monkeypatch, database_url):
-    monkeypatch.setenv("DATABASE_URL", database_url)
-    from space2stats.app import app
-
-    with TestClient(app) as test_client:
-        yield test_client
 
 
 def test_read_root(client):
